Refactor, using 'src' instead of 'bin' to hold the source code.
wee_resources becomes weewx_data, and is put in version control.
The build documentation, as well as the 'user' subdirectory, are now
in it.

This allows weewxd to be run directly from the git repository for
two reasons: 1) the documentation does not have to be built first.
2) because the source code is now under 'src', while the 'user'
subdirectory is in weewx_data, weewxd does not accidentally pick
up the repository's version of the 'user' subdirectory.